lesson 7/requirements.txt
mcp
fastmcp
langchain-openai
langchain-core
requests
crewai
python-dotenv
crewai-tools[mcp]
langchain-experimental
